DS-685: Move version selector components into addon directory
